// src/app/api/models/register-command.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ReligiousStatus } from '../models/religious-status';
import { UserRole } from '../models/user-role';
export interface RegisterCommand {
  email?: string;
  firstName?: string;
  lastName?: string;
  password?: string;
  religion?: ReligiousStatus;
  role?: UserRole;
}